refactor: split up http into separate modules
